third_party/blink/web_tests/platform/win/external/wpt/html/browsers/browsing-the-web/scroll-to-fragid/scroll-restore-on-reload-non-ascii-fragment-expected.txt

This is a testharness.js-based test.
[FAIL] Reload preserves scroll position for ASCII fragment
  assert_equals: scroll position should be preserved after reload, not scrolled to fragment expected 0 but got 608
[FAIL] Reload preserves scroll position for non-ASCII fragment
  assert_equals: scroll position should be preserved after reload, not scrolled to fragment expected 0 but got 1228
Harness: the test ran to completion.
